+++ sensu_plugin/plugin.py
@@ -1,30 +1,36 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-#
-# Copyright (C) 2014 - S. Zachariah Sprackett <zac@sprackett.com>
-#
-# Released under the same terms as Sensu (the MIT license); see LICENSE
-# for details.
 
 from __future__ import print_function
-import atexit
-import sys
+
 import argparse
+import atexit
 import os
+import sys
 import traceback
+
 from collections import namedtuple
+
 from sensu_plugin.exithook import ExitHook
 
+# create a namedtuple of all valid exit codes
 ExitCode = namedtuple('ExitCode', ['OK', 'WARNING', 'CRITICAL', 'UNKNOWN'])
 
 
 class SensuPlugin(object):
-    def __init__(self):
+    '''
+    Base class used by both checks and metrics plugins.
+    '''
+    def __init__(self, autorun=False):
+
         self.plugin_info = {
             'check_name': None,
             'message': None,
             'status': None
         }
+
+        # create a method for each of the exit codes
+        # and register as exiy functions
         self._hook = ExitHook()
         self._hook.hook()
 
@@ -34,18 +40,25 @@ class SensuPlugin(object):
 
         atexit.register(self.__exitfunction)
 
+        # Prepare command line arguments
         self.parser = argparse.ArgumentParser()
         if hasattr(self, 'setup'):
             self.setup()
         (self.options, self.remain) = self.parser.parse_known_args()
 
-        self.run()
+        if autorun:
+            self.run()
 
     def output(self, args):
-        print("SensuPlugin: %s" % ' '.join(str(a) for a in args))
+        '''
+        Print the output message.
+        '''
+        print("SensuPlugin: {}".format(' '.join(str(a) for a in args)))
 
     def __make_dynamic(self, method):
-
+        '''
+        Create a method for each of the exit codes.
+        '''
         def dynamic(*args):
             self.plugin_info['status'] = method
             if not args:
@@ -59,9 +72,16 @@ class SensuPlugin(object):
         setattr(self, dynamic.__name__, dynamic)
 
     def run(self):
+        '''
+        Method should be overwritten by inherited classes.
+        '''
         self.warning("Not implemented! You should override SensuPlugin.run()")
 
     def __exitfunction(self):
+        '''
+        Method called by exit hook, ensures that both an exit code and
+        output is supplied, also catches errors.
+        '''
         if self._hook.exit_code is None and self._hook.exception is None:
             print("Check did not exit! You should call an exit code method.")
             sys.stdout.flush()